A bunch of bugfixes and candidate solution for large problem

+++ Dijkstra/Dijkstra/large.h
@@ -3,6 +3,7 @@
 #include <string>
 #include <algorithm>
 #include <sstream>
+#include <vector>
 #include "common.h"
 
-bool SolveLargeProblem(double L, double X, std::string line);
+bool SolveLargeProblem(long L, long X, std::string line);
